\name{impulse}
\docType{data}
\alias{impulse}
\title{PK profile after drug intravenous administration}
\description{
This data set gives the time and concentration of the hypothetical drug after its intravenous administration. This is the simulated data set.
}
\usage{data(impulse)}
\format{matrix}
\keyword{datasets}
