\name{input}
\docType{data}
\alias{input}
\title{In vivo absorption of the drug}
\description{
This data set gives the time and cumulative amount of the hypothetical drug absorbed. It is also used as in vitro dissolution for Rivivc example of IVIVC level A. This is the simulated data set.
}
\usage{input}
\format{matrix}
\keyword{datasets}
