\name{Rivivc-package}
\alias{Rivivc-package}
\alias{Rivivc}
\docType{package}
\title{
IVIVC LEVEL A
}
\description{
This package performs linear iv vitro in vivo correlation of linear level A. It provides numerical convolution/deconvolution procedures with unequal time steps and no assumptions about the function shapes. 

}
\details{
\tabular{ll}{
Package: \tab Rivivc\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2012-10-03\cr
License: \tab GPLv3\cr
}
}
\author{
Aleksander Mendyk and Sebastian Polak

Maintainer: Aleksander Mendyk <mfmendyk@cyf-kr.edu.pl>

}
\references{
Langenbucher (2003) F. Handling of computational in vitro/in vivo correlation problems by Microsoft Excel: III. Convolution and deconvolution. \emph{Eur J Pharm Biopharm.} \bold{56}, 429-37.
}
\keyword{ package }
